check iframe and simply + refactor
