+++ src/components/Footer.jsx
@@ -1,6 +1,9 @@
 import { IconGitHub } from '../Icons'
+import { IS_IFRAME } from '../consts'
 
 export default function Footer () {
+  if (IS_IFRAME) return null
+
   return (
     <section className='fixed bottom-0 right-0 z-10 p-4'>
       <a
